Create zabbix.yaml
Before running this configuration, create the following folders in your Synology NAS:

/volume1/docker/zabbix/mysql for MySQL data.
/volume1/docker/zabbix/zabbix-server for Zabbix server files.
/volume1/docker/zabbix/zabbix-web for Zabbix web interface files.
To change the default port of the Zabbix web interface from 80 to 88 when using host network mode, you would typically need to modify the Nginx configuration within the Zabbix web container. However, since host networking is used and the container shares the host's network stack, the Nginx configuration file inside the container would need to be changed to listen on port 88 instead of port 80.